!!! Massive changes
Investigating the source of the massive error bars on angular.

Imported and adjusted C++/IGOR code into Python. => weave functions appears. They are duplicate of existing ones.

Conclusions: Source of the problem is how to treat the propagation of error with lstsq method. Otherwise they both are the same (Delta<1e-7)